Rename users queries file
